*refactoring of ConvertResizeNode.
