+++ prj/vs2022/CvtOnnx.vcxproj
@@ -96,6 +96,7 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertReduceSumNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertReluNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertReshapeNode.cpp" />
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertResizeNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertRoundNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertSignNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertSqueezeNode.cpp" />
