+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -279,6 +279,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertReshapeNode.cpp">
       <Filter>OnnxRuntime\R</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertResizeNode.cpp">
+      <Filter>OnnxRuntime\R</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\Cvt\OnnxRuntime\OnnxRuntime.h">
