+++ src/Cvt/OnnxRuntime/OnnxRuntime.h
@@ -61,90 +61,6 @@ namespace Synet
 
         //-----------------------------------------------------------------------------------------
 
-        bool ConvertResizeNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& original, LayerParam& layer)
-        {
-            if (layer.src().size() == 4)
-            {
-                const LayerParam * src1 = GetLayer(layers, layer.src()[1]);
-                if (src1->type() != Synet::LayerTypeConst || src1->weight()[0].dim()[0] != 0)
-                    return false;
-                const LayerParam * src2 = GetLayer(layers, layer.src()[2]);
-                if (src2->type() != Synet::LayerTypeConst || src2->weight()[0].dim()[0] != 0)
-                    return false;
-                layer.src().erase(layer.src().begin() + 1, layer.src().begin() + 3);
-                const LayerParam* src1b = GetLayer(layers, layer.src()[1]);
-                if (src1b->type() == Synet::LayerTypeConst)
-                {
-                    layer.weight() = src1b->weight();
-                    layer.src().resize(1);
-                }
-            }           
-            else if (layer.src().size() == 3)
-            {
-                const LayerParam * src1 = GetLayer(layers, layer.src()[1]);
-                if (src1->type() != Synet::LayerTypeConst || src1->weight()[0].dim()[0] != 0)
-                    return false;
-                layer.src().erase(layer.src().begin() + 1);
-                const LayerParam* src1b = GetLayer(layers, layer.src()[1]);
-                if (src1b->type() == Synet::LayerTypeConst)
-                {
-                    layer.weight() = src1b->weight();
-                    layer.src().resize(1);
-                }
-            }
-            if (layer.src().size() == 2)
-            {
-                const LayerParam * src1 = GetLayer(layers, layer.src()[1]);
-                if (src1->type() == Synet::LayerTypeMeta && src1->meta().type() == Synet::MetaTypeConst)
-                {
-                    const TensorParam & alpha = src1->meta().alpha();
-                    if (alpha.shape().size() == 1 && alpha.shape()[0] == 4)
-                    {
-                        layer.interp().height() = (int32_t)alpha.i64()[2];
-                        layer.interp().width() = (int32_t)alpha.i64()[3];
-                        layer.src().resize(1);
-                    }
-                    else
-                        return false;
-                }
-                else if (src1->type() == Synet::LayerTypeConst)
-                {
-                    layer.weight() = src1->weight();
-                    layer.src().resize(1);
-                }
-            }
-
-            String mode;
-            if (!ConvertAtrributeString(node, "mode", mode))
-                return false;
-            if (mode == "nearest")
-                layer.interp().interpolationType() = InterpolationTypeNearest;
-            else if (mode == "linear")
-                layer.interp().interpolationType() = InterpolationTypeBilinear;
-            else
-                return false;
-
-            if (GetAtrribute(node, "coordinate_transformation_mode"))
-            {
-                String coordTransf;
-                if (!ConvertAtrributeString(node, "coordinate_transformation_mode", coordTransf))
-                    return false;
-                if (coordTransf == "pytorch_half_pixel")
-                    layer.interp().coordinateTransformType() = CoordinateTransformTypeHalfPixel;
-                else if (coordTransf == "asymmetric")
-                    layer.interp().coordinateTransformType() = CoordinateTransformTypePytorch;
-                else if (coordTransf == "half_pixel")
-                    layer.interp().coordinateTransformType() = CoordinateTransformTypeHalfPixel;
-                else if (coordTransf == "align_corners")
-                    layer.interp().coordinateTransformType() = CoordinateTransformTypeCaffe;
-                else
-                    return false;
-            }
-
-            layer.type() = Synet::LayerTypeInterp;
-            return true;
-        }
-
         bool ConvertScaledDotProductAttentionNode(const onnx::NodeProto& node, const LayerParams& layers, LayerParam& layer, Bytes& reordered)
         {
             if (!CheckSourceNumber(layer, 4))
